% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{build_report}
\usage{
build_report(file, outfile = NULL, type = c("word_document",
  "pdf_document"), ...)
}
\arguments{
\item{file}{A path to the RMarkdown file (required).}

\item{outfile}{The path of the built document. Defaults to the working
directory.}

\item{type}{The kind of output document; either a Microsoft Word (.docx)
or Acrobat Portable Document Format (.pdf) document. Defaults to the
former.}

\item{...}{Arguments passed on to \code{rmarkdown::render()}.}
}
\value{
This function has no return value. The result of the operation is
    a built document in the specified file format.
}
\description{
build_report
}
