% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_webreport}
\alias{build_webreport}
\title{Build a report from a source file}
\usage{
build_webreport(data.source, outfile = NULL, ...)
}
\arguments{
\item{data.source}{A database file containing web data}

\item{outfile}{The path of the built document. Defaults to the working
directory.}

\item{...}{Additional arguments passed to \code{rmarkdown::render()}.}
}
\value{
This function has no return value. The result of the operation is
a built document in the specified file format.
}
\description{
Creates a report from an R Markdown template file that is
found in the package (precisely in \code{inst/rmarkdow}) and utilises a
data previously stored in a local database.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
