% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{chart_tweets}
\alias{chart_tweets}
\title{chart_tweets}
\usage{
chart_tweets(tbl)
}
\arguments{
\item{tbl}{A table; principally object returned by \code{compare_mentions}}
}
\description{
Generates a barplot that compares the propotion of tweets with differnt
search terms, colours it and gives it appropriate labels
}
