% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{choose_insight}
\alias{choose_insight}
\title{Collect data on Insights from a Facebook Page.}
\usage{
choose_insight(type = c("page_fan_adds", "page_fan_removes",
  "page_views_login", "page_views_logout", "page_views", "page_story_adds",
  "page_impressions", "page_posts_impressions", "page_consumptions",
  "post_consumptions_by_type", "page_fans_country"))
}
\arguments{
\item{type}{The type of Facebook insight being queried.}
}
\description{
Retrieves and preprocesses insights from NESREA Facebook Page
}
\note{
This is a wrapper for \code{Rfacebook::getInsights}
}
