% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{choose_platform}
\alias{choose_platform}
\title{Select a social media network}
\usage{
choose_platform(site)
}
\arguments{
\item{site}{A character vector; a social media site (supported sites are
Facebook and Twitter.}
}
\value{
Returns an integer value, 1 for Twitter and 2 for Facebook, which
is used internally for indexing other relevant functions.
}
\description{
Selects the appropriate social media platform being analysed
}
\examples{
choose_platform("Facebook")
}
