% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{collect_tweets}
\alias{collect_tweets}
\title{Collect Some Tweets}
\usage{
collect_tweets(string)
}
\arguments{
\item{string}{A character vector of length 1 of a search term}
}
\description{
Collects tweets containing a particular term, to a maximum of 1,000.
}
