% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{compare_mentions}
\alias{compare_mentions}
\title{Comparing Twitter Mentions}
\usage{
compare_mentions(terms, n = 500L, from = NULL, to = NULL)
}
\arguments{
\item{terms}{Character vector of Twitter handles of length > 1}

\item{n}{Maximum number of tweets to download (default is 500)}

\item{from}{Tweets from a given data in the format YYYY-MM-DD}

\item{to}{Tweets up to a given data in the format YYYY-MM-DD}
}
\description{
Prints a proportions table of number of tweets for various search terms.
}
