% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{compute_emotional_valence}
\alias{compute_emotional_valence}
\title{Compute the emotional valence in a given text}
\usage{
compute_emotional_valence(text.var)
}
\arguments{
\item{text.var}{A character vector or an object coercible to one.}
}
\description{
Calclates the polarity for each text field
}
\note{
This operation ought to generate series of warnings because
of multiple punctuations that expectedly occur in the text, but
this has been suppressed.
}
