% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dens.R
\name{dens_plot}
\alias{dens_plot}
\title{Draw a density plot of social data}
\usage{
dens_plot(data, platform)
}
\arguments{
\item{data}{An object of class \code{data.frame}}

\item{platform}{A character string representing one of 2 social media sites
-- \href{http://twitter.com/}{Twitter} or
\href{htpp://facebook.com/}{Facebook}}
}
\value{
A \code{ggplotObj} object which upon printing draws a density plot.
}
\description{
Draw a Kernel density plot from previously stored social media data
}
