% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{display_twts}
\alias{display_twts}
\title{display_twts}
\usage{
display_twts(x)
}
\arguments{
\item{x}{A data frame created from a list of Twitter JSON objects}
}
\description{
Displays a density plot of tweets
}
