% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_all_data}
\alias{download_all_data}
\title{download_all_data}
\usage{
download_all_data(data.store)
}
\arguments{
\item{data.store}{An SQLite database}
}
\description{
Download various web data from the internet
}
