% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-fbposts.R
\name{download_fb}
\alias{download_fb}
\title{Download Facebook Page posts}
\usage{
download_fb(keyword, dBase)
}
\arguments{
\item{keyword}{A character vector containing a search term}

\item{dBase}{A database file}
}
\description{
Dowwload data from a Facebook page containing a particular term.
}
