% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{fbTokenObj}
\alias{fbTokenObj}
\title{fbTokenObj}
\usage{
fbTokenObj(app_id, app_secret)
}
\arguments{
\item{app_id}{The Facebook App ID}

\item{app_secret}{The Facebook App Secret}
}
\value{
An S3 Object containing a generated Facebook access token and the
token's expiration date.
}
\description{
Constructor an S3 object containing an access token and its
expiry date
}
\details{
The key thing about this function is that we want to be able to
keep in store the expected expiry date the token according to the
prevailing Facebook API policy, so that on loading it we can
confirm whether it is still valid or not.
}
