% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{fetch_token}
\alias{fetch_token}
\title{fetch_token}
\usage{
fetch_token()
}
\description{
fetch_token
}
\details{
Checks for local storage of the access token. If it is already present
we also check whether it has expired or not, in line with Facebook's
policy on token changes. The App credentials used (App Id & App Secret)
are as available via the App dashboard,
}
