% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{generate_wordcloud}
\alias{generate_wordcloud}
\title{Generate a Tag Cloud}
\usage{
generate_wordcloud(data, pol.list, site)
}
\arguments{
\item{data}{An object of class \code{data.frame} that has a column
with computed emotional valence}

\item{pol.list}{A list of polarities}

\item{site}{A character vector of length 1; name of the social media
platform}
}
\description{
Generates a tag cloud
}
