% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{make_corpus}
\alias{make_corpus}
\title{make_corpus}
\usage{
make_corpus(GText, stem = TRUE)
}
\arguments{
\item{GText}{A column of text taken from the data}

\item{stem}{Whether stemming will be carried out.}
}
\description{
Makes a volatile corpus and prepares text
}
