% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{make_word_table}
\alias{make_word_table}
\title{Create a word table}
\usage{
make_word_table(pol.list)
}
\arguments{
\item{pol.list}{A list of polarities}
}
\description{
Makes a table of words on eiher side of the emotional spectrum i.e.
positive/negative
}
