% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Downloaded Facebook Data for further Processing.}
\usage{
prepare_data(df)
}
\arguments{
\item{df}{An object of class \code{data.frame}, specifically downloaded
via that Facebook API.}
}
\description{
Processes Facebook data prior to analysis
}
