% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{process_stored_tweets}
\alias{process_stored_tweets}
\title{Process data types of previously stored tweets}
\usage{
process_stored_tweets(data)
}
\arguments{
\item{data}{An object of class \code{data.frame} specifically containing
tweets stored in a local SQLite database.}
}
\description{
Process data types of previously stored tweets
}
\note{
When the database is queried with \code{RSQLIte::dbReadTable()},
the data types of some of the columns do not conform to the desired state.
This function changes the column with dates to POSIX format, while
appropriate colums of the data frame are converted to \code{logical}.
}
