% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{remove_nonreadables}
\alias{remove_nonreadables}
\title{Remove Characters Not Readable by Humans}
\usage{
remove_nonreadables(string = NULL)
}
\arguments{
\item{string}{A character vector of length 1}
}
\description{
Makes sure that text-based columns contain human readable
characters only
}
