% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{renew_fb_cred}
\alias{renew_fb_cred}
\title{Renew Facebook Token}
\usage{
renew_fb_cred()
}
\description{
Obtain a fresh Facebook access token upon expiry of the extant one
}
\details{
When run, the Facebook authentication process is initiated, which
involves copying and pasting a URL into the relevant field in the App's
'Settings' page. For more information on Facebook authentication see
\code{\link[Rfacebook]{fbOAuth}}.
}
\note{
The two credentials that are used to obtain the token (i.e. the App ID
and the App secret) are specific to NESREA. Only users with access to the
App's 'Settings' page at \url{https://developers.facebook.com/} can effect
the renewal.
}
