% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{return_text}
\alias{return_text}
\title{return_text}
\usage{
return_text(df, metric)
}
\arguments{
\item{df}{An object of class 'data.frame'}

\item{metric}{A social media metric to be explored}
}
\description{
Returns a message that matches a particular metric (used
only inside the body text)
}
