% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-website.R
\name{scrape_items}
\alias{scrape_items}
\title{scrape_items}
\usage{
scrape_items(page, selector, verbose = FALSE)
}
\arguments{
\item{page}{A web page in class \code{xml_document} and \code{xml_node}.}

\item{selector}{A CSS selector on the web page; used to identify data of
interest.}

\item{verbose}{Logical; whether to print out informative messages.}
}
\description{
Convenienly collect data based on CSS selector
and return it as a character vector
}
