% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{show_datasets}
\alias{show_datasets}
\title{Display Datasets}
\usage{
show_datasets(dbfile)
}
\arguments{
\item{dbfile}{A path to a database file.}
}
\description{
Present the data for inspection
}
\details{
This function is a enables the display of all the tables in the
database in the spreadsheet-like format prrovided by \link{[utils]{View}}.
}
