% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{show_tweets_containing}
\alias{show_tweets_containing}
\title{show_tweets_containing}
\usage{
show_tweets_containing(word)
}
\arguments{
\item{word}{A character vector of strings that are being searched for}
}
\description{
Search for and display tweet(s) containing a particular word
}
