% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{store_post_details}
\alias{store_post_details}
\title{store_post_details}
\usage{
store_post_details(conn, data)
}
\arguments{
\item{conn}{- an SQLite database connection}

\item{data}{- a dataframe returned by \code{Rfacebook::getPage}}
}
\description{
Collect data frame of post details
}
\note{
Some posts e.g. 'Events', do not come in desired format,
so there's a need to condition on that possibility
}
