% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{token_expiry}
\alias{token_expiry}
\title{Check Access Token Validity}
\usage{
token_expiry()
}
\description{
Check the expiry date of a Facebook access token.
}
\details{
The S3 object \code{fbTokenObj} has a member that represents the
date the expected expiry date as determined by the exisiting Facebook
policy for access tokens.
}
\examples{
inherits(token_expiry(), 'Date')

}
