% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tweets.R
\name{update_db}
\alias{update_db}
\title{Update a local database with Twitter data}
\usage{
update_db(dBase)
}
\arguments{
\item{dBase}{A database file for storage of downloaded tweets.}
}
\description{
Update a local database with Twitter data
}
