% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{visualise_pol_diff}
\alias{visualise_pol_diff}
\title{Visualise Different Polarities}
\usage{
visualise_pol_diff(pol.list)
}
\arguments{
\item{pol.list}{A list of polarities}
}
\description{
Displays the occurence of words on either side of the
spectrum using a dot plot.
}
