% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw-functions.R
\name{compare_mentions}
\alias{compare_mentions}
\title{Comparing Twitter Mentions}
\usage{
compare_mentions(x, n = 50L)
}
\arguments{
\item{x}{character vector of search terms}

\item{n}{max. number of tweets to download (default is 50)}
}
\description{
Prints a proportions table of number of tweets for various search terms.
}
