% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_all_data}
\alias{download_all_data}
\title{Download Data in General}
\usage{
download_all_data(keyword, data.store)
}
\arguments{
\item{keyword}{A character vector containing a search term}

\item{data.store}{An SQLite database}
}
\description{
Downloads data from all the relevant website and social media platforms.
}
