% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tweets.R
\name{download_tweets}
\alias{download_tweets}
\title{Download tweets and store in a local database}
\usage{
download_tweets(keyword, db)
}
\arguments{
\item{keyword}{A character vector containing a search term}

\item{db}{A local database file}
}
\description{
Download tweets and store in a local database
}
\note{
This function currently only works with SQLite databases
}
