% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-website.R
\name{download_website}
\alias{download_website}
\title{Download Website Data}
\usage{
download_website(database)
}
\arguments{
\item{database}{A database file for storing downloaded web data}
}
\description{
Downloads the webpage as an XML document and then makes a
dataframe from vectors of text scraped from parts of the page via CSS
selectors.
}
