% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{make_ts}
\alias{make_ts}
\title{Plot Weekly Time-Series Data For Online Platforms}
\usage{
make_ts(data, columnname, colour = "blue")
}
\arguments{
\item{data}{An object of class \code{data.frame}}

\item{columnname}{A character string representing Date-Time values, or values
that can be coerced to Date-Time values.}

\item{colour}{A character string giving a valid R built-in colour as
represented in \code{\link[grDevices]{colors}}.}
}
\value{
A base time-series plot.
}
\description{
Uses locally stored data from Website, Facebook and Twitter profiles to plot
time-series for the rolling year in weekly time units.
}
\details{
The data used for this function i.e. the date variable, is expected
to be either a nummeric vector (as stored by Twitter and Website data) or as
a character vector (Facebook). The Facebook date elements are stored in the
POSIX-compatible \code{YYYY-MM-DDTHH:MM:SS+} format. Any other presentation
produce an error.
}
