% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{make_ts}
\alias{make_ts}
\title{Plot Weekly Time-Series Data For Online Platforms}
\usage{
make_ts(data, platform, ...)
}
\arguments{
\item{data}{An object of class \code{data.frame}}

\item{platform}{A character string, with the name of the chosein online
platform. Options are Facebook, Twitter and Website.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
Has no return value per se. It results in the production of a
time-series plot.
}
\description{
Uses locally stored data from Website, Facebook and Twitter profiles to plot
time-series for the rolling year in weekly time units.
}
