% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-function-prototypes.R
\name{plain_dens_plot}
\alias{plain_dens_plot}
\title{plain_dens_plot}
\usage{
plain_dens_plot(data, platform)
}
\arguments{
\item{data}{An object of class 'data.frame'}

\item{platform}{A character vector of length 1 with the name of the social
media site}
}
\description{
Plots a simple density plot for the data
}
