% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{renew_fb_cred}
\alias{renew_fb_cred}
\title{Renew Facebook Token}
\usage{
renew_fb_cred()
}
\description{
Obtain a fresh Facebook access token upon expiry of extant one
}
\note{
The app id and app secret that are used to obtain the token are
particular to NESREA.
}
